pyproject.toml
[tool.poetry]
name = "backend"
version = "0.1.0"
description = ""
authors = ["Jason Kringen <jason.kringen@gmail.com>"]
readme = "README.md"
package-mode = false

[tool.poetry.dependencies]
python = "^3.12"
django = "^5.2.7"
djangorestframework = "^3.16.1"
djangorestframework-simplejwt = "^5.5.1"
django-cors-headers = "^4.9.0"
pydantic = {extras = ["email"], version = "^2.12.4"}
uvicorn = "^0.38.0"
psycopg2 = "^2.9.11"
whitenoise = "^6.11.0"

[tool.poetry.group.dev.dependencies]
pytest = "^8.4.2"
ruff = "^0.14.3"
faker = "^37.12.0"

[tool.ruff.lint.per-file-ignores]
"seed_db.py" = ["E402"]

[tool.ruff]
line-length    = 120
target-version = "py312"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
